requirements.txt
# Core dependencies
ultralytics>=8.1.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0

# Optional but recommended
opencv-contrib-python>=4.8.0
